Pax format now instead of ustar format
HDR_LENGTH=(512)*3 according to my analysis. That is, for each
file/directory/link, there is such header precedes the data section.
This is obviously quite a bit overhead. One thing we can do is to have a
command line option to decide if PAX format is a must. However, as we
are targeting Lustre environment, let's make it a default at this point.