Made option name and help string consistent in m4/lx_find_dtcmp.m4
m4/lx_find_dtcmp.m4 caused configure to document "--with-dtcmp", but
the option that configure processed was in fact called
"--with-libdtcmp".  This revision renames the option to match the help
string.

Lots of other files are updated as well as the result of running
autoreconf on the directory.  Those are all machine-generated changes,
not human edits.