Add Lustre grouplock support in dcp2
Add a new option "-g/--groulock $GID" to dcp2. When opening a file
on Lustre, grouplock will be obtained automatically. This would
not only imporve the performance of reading/writing files in
Lustre file system, but also prevents other processes from
changing the files being copied. The grouplock ioctl on non-Lustre
file system will fail, but the failure will be ignored so won't
cause any problem.

Also, an new option --enable-lustre is added to configure system.
The Lustre support will be only enabled when this option is enabled,
and at the same time the head files of Lustre exist.

Change-Id: I9882afe08e66ce7b006412db3c5e88dee9c9be73
Signed-off-by: Li Xi <lixi@ddn.com>