Add seperate dsync tool extracted from dcmp
This is still very much a work in progress. There is a lot of
duplicated code that exists in both dsync and dcmp, so work needs
to be done to extract the common code and make it available to both
tools. The dsync tool is functional, but needs heavy refactoring.