dwalk: add default file histogram
When file_histogram option is used this creates a file histogram
of all walked files. This option does not require any range inputs.
It will give the user a count of files in each bin. The bins
are created based on the max file size. So if the Max file size
is between 2^20 - 2^30, then this is the last bin. The bins
start with the first bin for 0 byte files, then goes up from
2^10, 2^20, 2^30, etc. The bins are created dynamically based
on the max file size.

Signed-off-by: Danielle Sikich <sikich1@llnl.gov>