dfind: fix -type to allow all file types
The check for MFU_PRED_TYPE() incorrectly used the file type as a
mask for the file mode, which would lead to collisions between types
e.g. S_IFREG (0100000) and S_IFLNK (0120000) or S_IFSOCK (0140000).
This would result in "-type f" also finding symlinks.

Properly use S_IFMT as the type mask for the file, and simplify the
return type since it is already a boolean when computed.

Update the man page to report -type can check for all file types.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>