add daos-preserve option for preserving daos metadata
This option depends on the hdf5 library, and can only be used
when moving data to or from a POSIX filesystem. If a user
moves data from a POSIX DAOS Container to NFS or Lustre they
can write the container properties and user attributes to
an hdf5 file. When a user is ready to move their data back
into a container they can pass the metadata file into the
copy and the properties will be read from the hdf5 file.
The container properties and user attributes are set on
container creation in the case the the data is moving from
a POSIX filesystem to a DAOS POSIX container.

If a user is moving data from one DAOS container to another
the metadata is already preserved from the source to destination,
and passing the option is not necessary.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>