serialization optimizations for DAOS_OF_KV_FLAT (#482)
* add daos-gen tool to generate a DAOS_OF_KV_FLAT objects
    * skip akey iteration on serialize and deserialize for DAOS_OF_KV_FLAT

If an object has the DAOS_OF_KV_FLAT feature set, then there is no
need to do any sort of akey iteration. The daos_kv API is used
because all that is needed is oid, dkey, and the value. The
obj API and the daos_kv API cannot be mixed, or it will cause
errors. The KV value is stored inside of the dkey dataset.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>