daos: update setting of container label
Remove copying of container label when copying from
DAOS->DAOS. The container label in DAOS is now required
to be unique.

The label is serialized, but only deserialized if it does not
exist.

Fix bug in cont_deserialize_all_props where a copy of the
pointer was being passed to it instead of the address to
the pointer, causing the properties to not be set correctly
in cont_create.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>