daos: negative test case bug fixes
This fixes bugs found in a test case that is currently being
skippin the DAOS CI. There were a couple of regressions that
need fixing. If a daos source string without a container was
passed in dcp was not reporting invalid arguments were passed.
Also, it should not matter if the source/dest pool or cont UUID
are upper or lower case. It was incorrectly reporting that the
DAOS source and destination were the same in that case.

    * properly check if a source pool and source cont are passed
    * allow passing in upper or lower case UUIDs for src and dst
    * return -1 when daos read or write fails

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>