fix bufsize typo in mfu_copy_xattrs and DCOPY_copy_xattrs
Initial buffer allocated is size 1204.  The intended size was 1024,
indicated in the commit message (below).  1024 fits better in PAGE_SIZE
and so may slightly improve memory management.

The typo was originally introduced into DCOPY_copy_xattrs(), then the
code was copied into mfu_copy_xattrs() as part of later work.

The original commit that introduced the typo was:

	commit 9a211b5f48f322d439b6f03dda6009dbff78e4c3
	Author: Adam Moody <moody20@llnl.gov>
	Date:   Thu Nov 21 14:19:59 2013 -0800

	    start out with 1024 byte buffer for xattrs

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>