test_walkfail.sh: implement src walk fail test
Implement a test where dsync --delete is run, and the walk of the
source tree is caused to fail by making a directory unreadable.

dsync currently has a bug which causes it to fail that test,
https://github.com/hpc/mpifileutils/issues/587

After the source walk fails, dsync continues with the dest walk
followed by the copies that the deletes.

Because the source walk failed, the list of files and directories
on the source is incomplete. Files or directories that are already
on both source and destination, but which were not walked on the
source, are missing from the source list.  Those files or directories
are then deleted from the destination tree.