dsync: disable --delete on src walk error
Record the ocurrence of errors in mfu_flist_walk_param_paths() and
friends.  Alter mfu_flist_walk_path() and friends to return an int
indicating success (0) or failure.

Add (void) to mfu_flist_walk_param_paths() calls where we do not use the
return value.

In dsync, if there was an error during the walk of the source tree, the
source flist generated is likely incomplete.

In the case where the source and destination have the same files and
directories before dsync --delete is run, an incomplete source flist
results in files being incorrectly deleted from the destination.

When an error was reported for the source walk, warn the user and
disable the --delete option.  Continue with the dsync, so that any
new files identified are still copied and metadata is synced.

See rsync(1) for the --delete option, it takes the same approach.

Fixes #587

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>