dsync: disable delete on src walk error
If there is an error during the walk of the source tree, the source flist
generated is likely incomplete.

In the case where the source and destination have the same files and
directories before dsync --delete is run, an incomplete source flist
results in files being incorrectly deleted from the destination.

When an error is detected on the source walk, warn the user and disable
the delete option.

See rsync(1) for the --delete option, it takes the same approach.