dfilemaker: add parent_path positional argument
Instead of creating files in the current working directory by default,
require the user to provide a path to the parent directory under which
files, directories, and links will be created.

This makes it explicit what file system will be affected.

It's a frequent issue on LC computers that users accidentally run
parallel tasks on NFS file systems which aren't able to handle the load,
degrading service for everyone.

It's also sometimes the case that users run a utility without any
arguments expecting a usage message.  With the old behavior, that would
result in a large directory tree being created in a directory the user
didn't intend.

This change addresses both of these issues.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>