dfilemaker: pass correct pointer to mfu_free()
Commit
	commit fc70c2ce3d7499685697157db91ccd645f2b031e
	Author: Olaf Faaland <faaland1@llnl.gov>
	Date:   Thu Nov 14 15:44:34 2024 -0800

	dfilemaker: address review comments for PR #269

introduced an error.  mfu_free() takes a pointer to a pointer,
because it both frees the memory, and also sets the pointer to
NULL to prevent double-free bugs.

The above commit passed the address of the memory to be freed,
not the address of the pointer.  This meant both that the
wrong memory was freed, potentially, and also that the code
attempted to set some arbitrary set of bytes to NULL.

In testing, this triggered a segfault when running dfilemaker.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>