dfilemaker: fail and stop execution on unrecognized option
If getopt_long() encounters an unknown option, print usage information
and stop execution.

Continuing to process options and create files, after encountering an
unknown option, will create a file tree different than the one the user
intended.  For example, if the user mistypes the option for file size,
e.g.
	"--sizeRRR=2MB-10000GB"
instead of
	 "--size=2MB-10000GB"

proceeding would mean we use the default file size range when
constructing the tree instead of the one the user intended.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>