build: promote dfilemaker and retire dfilemaker1
Retire dfilemaker1, a tool which creates some number of files in the
current working directory and does not create a tree, store a variety of
data, or generally produce an interesting or varied set.

Remove dfilemaker from experimental, to replace dfilemaker1.  dfilemaker
supports creation of trees of various sizes, numbers of levels, and data
content.

Remove mention of unimplemented options width and ratio, as when they
will be implemented is unknown.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>