test_dsync: make first arg MFU binary directory
Instead of passing the path to individual mfu binaries, e.g. dsync and
dfilemaker, pass the path to the directory in which they reside.

Then one argument is enough to point to them all.  Testing with some MFU
components from one build, and some from a different build, is not a
normal test scenario, so no need to make scripts more complex to support
it.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>