test_multiple_syncs.sh: verify sequential syncs change nothing
Test multiple dsyncs, and test dsync against rsync.

1. Running dsync twice, the second time no changes should be reported.
2. Running rsync followed by dsync, the dsync should report no changes.
3. Running dsync followed by rsync, the rsync should report no changes.

rsync is run with -av -HAX to copy hard links, ACLs, and xattrs.

However, the test data copied is created with dfilemaker, which does not
create ACLs or xattrs, or named pipes, and may not create files with
multiple hard links.  This should be improved.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>