Fix for issue dsync not updating times correctly #560.
Some file system do not support nanosecond part of the timespec for
mtime (last modification time). We were comparing the mtime regardless
of the support to mtime nanosecond. This was causing dsync to add files
to the copy to destination list even if the file were identical.

We added a check for source dir and destination dir to verify if the
timepsec nanosecond (tv_nsec) was set to 0. In which case we do not
compare the mtime nanosecond part of it (tv_nsec).

Cleaning up dead code left in the source files.
dsync.c and mfu_flist_copy.c

Cleaned up code of unused code and reverted src/common/mfu_flist_copy.c
The change in mfu_flist_copy.c is for another MR.