mfu: index hardlinks during tree walk
During tree walk with details, regular files with more than one nlink
are temporarily placed in a hardlinks flist. This flist is then globally
ordered by names and ranked to select one reference path per inode, and
flag all other paths to this inodes as hardlinks. The sorted hardlinks
flist is finally merged in global flist with all other items.

The paths name ordering is performed to ensure reproducibility between
two similar trees, thus minimizing the differences for dcmp and dsync
eventually.

This commit introduces a new structure inodes_hardlink_map_t used to
temporarily associate paths to inodes in reference/hardlinks solving
logic.

The type elem_t receives 2 new members: nlink, the number of links on an
inode, and ref, the reference path to this inode. The ref is NULL except
on hardlinks.

This commit also introduces a new filetype MFU_TYPE_HARDLINK, which is
used to distinguish hardlinks to inodes from reference paths which have
MFU_TYPE_FILE type.

The packed flist element now contains the filetype, even when details
are enabled, as there is now way to determine if an element is a regular
file or a hardlink based on stat result.

New functions mfu_[un]pack_sized_str() are introduced to manage packing
and unpackaging of optional strings with maximum length.

Signed-off-by: Rémi Palancher <remi@rackslab.io>