dcmp: support hardlinks
Add support of hardlinks in dcmp. The reference paths of hardlinks in
source and destination are compared. If not equal, strmap is updated to
flag them as different.

The branchs in items comparison logic is now based on filetype recorded
in flist rather than the file mode as there is no way to distinguish
reference paths and hardlinks with just the mode, both are regular
files.

Signed-off-by: Rémi Palancher <remi@rackslab.io>