dsync: support hardlinks
Add support of hardlinks in dsync. The reference paths of hardlinks in
source and destination are compared. If not equal, strmap is updated to
flag them as different.

The branchs in items comparison logic is now based on filetype recorded
in flist rather than the file mode as there is no way to distinguish
reference paths and hardlinks with just the mode, both are regular
files.

Additional logic is added with dsync_remove_hardlinks_with_removed_ref()
function to detect hardlinks whose references paths are marked for
deletion in destination. In this case, all the hardlinks pointing to
this reference are also marked for being replaced to avoid residual
links pointing to wrong inodes.

Signed-off-by: Rémi Palancher <remi@rackslab.io>