tests: automatic functional tests
This command adds many functional tests of dcmp, dcp, dsync, dtar and
dwalk, executed and automatically validated with Python standard
unittest library. This is designed to be easy to execute and integrate
in continuous integration systems.

Set two environment variables to define respectively the path to
mpifileutils binaries and arguments provided to mpirun, eg:

  $ export MFU_BIN=~/dev/bin
  $ export MFU_MPIRUN_ARGS="--bind-to none --oversubscribe -N 4"

And run all the tests:

  $ python3 -m unittest discover -v test

Or:

  $ pytest  # require pytest

The suite has utilities to check similarity between two trees, with the
possibility to specific paths and attributes (eg. mtime). It is also
possible to assert specific command outputs.

Most tests are run against a specific testing file tree to cover many
cases. Other tests are run with a file tree generated by dfilemaker.

Signed-off-by: Rémi Palancher <remi@rackslab.io>