dtar: fix creation with symlink and --preserve-*
Fix creation of corrupted archive with symlink when
--preserve-{xattrs,acl,flags} command line option is used.

When it is provided a valid file descriptor, libarchive
archive_read_disk_entry_from_file() calls fstat() which follows
symlinks. This commit introduces an exception for symlinks to not
provide a file descriptor to archive_read_disk_entry_from_file() which
makes it call lstat() on entry path and create a valid symlink archive
entry eventually.

fix #628

Signed-off-by: Rémi Palancher <remi@rackslab.io>