ignore FS_IOC_FIEMAP failure with ENOTSUPP
Lustre 2.15.6 can return ENOTSUPP (524) in response to an FS_IOC_FIEMAP
ioctl() call, for a file striped a certain way.  This is a bug, it
should return EOPNOTSUPP (95).

The file can still be copied, fiemap just can't tell us where the holes
are.

A brief survey using google searches suggests ENOTSUPP (524) was
historically intended for use internal to the kernel or drivers, but has
occasionally been returned to users by different software, in ways that
have meaning similar to ENOTSUP.

Ignore the error as with ENOTSUP (95) so the file is copied without the
use of fiemap.

Notify the user that the copied file is not sparse.

Fixes #644

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>