dsync: skip dst hardlink deletion when not in src
When the content of an hardlink reference is different in destination,
dsync iterates over the other hardlinks to mark them with different
content as well. In this loop, we must skip the hardlink which do not
have corresponding entry in source tree. They are handled in
dsync_only_dst() when --delete is set eventually.

Signed-off-by: Rémi Palancher <remi@rackslab.io>